# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/public/onboarding-leads": {
      "post": {
        "operationId": "PublicOnboardingLeadController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOnboardingLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingLeadDto"
                }
              }
            }
          }
        },
        "tags": [
          "Onboarding leads"
        ]
      }
    },
    "/api/v1/organizations/onboarding-leads": {
      "get": {
        "operationId": "OrganizationOnboardingLeadController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OnboardingLeadDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Onboarding leads"
        ]
      }
    },
    "/api/v1/organizations/onboarding-leads/{id}": {
      "patch": {
        "operationId": "OrganizationOnboardingLeadController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOnboardingLeadStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingLeadDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Onboarding leads"
        ]
      }
    },
...
    "/api/v1/enterprise/org": {
      "get": {
        "operationId": "EnterpriseOrganizationController_getOrg",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseOrgDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Enterprise organization"
        ]
      }
    },
    "/api/v1/enterprise/departments": {
      "post": {
        "operationId": "EnterpriseOrganizationController_createDepartment",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnterpriseDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDepartmentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Enterprise organization"
        ]
      }
    },
    "/api/v1/enterprise/locations": {
      "put": {
        "operationId": "EnterpriseOrganizationController_replaceLocations",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceEnterpriseLocationsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Enterprise organization"
        ]
      }
    },
    "/api/v1/enterprise/members": {
      "post": {
        "operationId": "EnterpriseOrganizationController_createMember",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnterpriseMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Enterprise organization"
        ]
      }
    },
    "/api/v1/enterprise/members/{membershipId}/disable": {
      "post": {
        "operationId": "EnterpriseOrganizationController_disableMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Enterprise organization"
        ]
      }
    },
    "/api/v1/government/members": {
      "get": {
        "operationId": "GovernmentOrganizationController_listMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GovernmentMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Government organization"
        ]
      },
      "post": {
        "operationId": "GovernmentOrganizationController_createMember",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGovernmentMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernmentMemberDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Government organization"
        ]
      }
    },
    "/api/v1/government/members/{membershipId}/disable": {
      "post": {
        "operationId": "GovernmentOrganizationController_disableMember",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Government organization"
        ]
      }
    },
    "/api/v1/platform/public-academy/campuses": {
      "get": {
        "operationId": "PublicAcademyController_listCampuses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicAcademyCampusDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Public academy"
        ]
      }
    },
    "/api/v1/platform/public-academy/campuses/{campusId}": {
      "patch": {
        "operationId": "PublicAcademyController_setCampusStatus",
        "parameters": [
          {
            "name": "campusId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPublicAcademyCampusStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicAcademyCampusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Public academy"
        ]
      }
    }
  },
  "info": {
    "title": "Zhixing Hub API",
    "description": "知行工坊模块化单体业务接口",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zhixing_session"
      }
    },
    "schemas": {
...
      "SubmitOnboardingLeadDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "UNIVERSITY",
              "ENTERPRISE"
            ]
          },
          "institutionName": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactMethod": {
            "type": "string",
            "description": "联系方式（电话 / 邮箱等）"
          },
          "intent": {
            "type": "string"
          },
          "captchaToken": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "institutionName",
          "contactName",
          "contactMethod",
          "intent",
          "captchaToken"
        ]
      },
      "OnboardingLeadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "UNIVERSITY",
              "ENTERPRISE"
            ]
          },
          "institutionName": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactMethod": {
            "type": "string"
          },
          "intent": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "OPENED",
              "CLOSED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "statusChangedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "kind",
          "institutionName",
          "contactName",
          "contactMethod",
          "intent",
          "status",
          "createdAt",
          "statusChangedAt"
        ]
      },
      "UpdateOnboardingLeadStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "OPENED",
              "CLOSED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "EnterpriseLocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "divisionCode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "divisionCode"
        ]
      },
      "EnterpriseDepartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
      "EnterpriseMemberDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "ENTERPRISE_ADMIN",
              "HR",
              "PROJECT_LEAD",
              "MENTOR",
              "ENTERPRISE_DASHBOARD"
            ]
          },
          "departmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "PENDING_ACTIVATION",
              "ACTIVE",
              "SUSPENDED"
            ]
          },
          "membershipStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "accountId",
          "membershipId",
          "displayName",
          "phone",
          "role",
          "departmentId",
          "accountStatus",
          "membershipStatus"
        ]
      },
      "EnterpriseOrgDto": {
        "type": "object",
        "properties": {
          "natureTagCode": {
            "type": "string"
          },
          "industryCategoryCode": {
            "type": "string"
          },
          "locations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnterpriseLocationDto"
            }
          },
          "departments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnterpriseDepartmentDto"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnterpriseMemberDto"
            }
          }
        },
        "required": [
          "natureTagCode",
          "industryCategoryCode",
          "locations",
          "departments",
          "members"
        ]
      },
      "CreateEnterpriseDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ReplaceEnterpriseLocationsDto": {
        "type": "object",
        "properties": {
          "locationCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "locationCodes"
        ]
      },
      "CreateEnterpriseMemberDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          },
          "role": {
            "type": "string",
            "enum": [
              "HR",
              "PROJECT_LEAD",
              "MENTOR",
              "ENTERPRISE_DASHBOARD"
            ]
          },
          "departmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "displayName",
          "phone",
          "role",
          "departmentId"
        ]
      },
      "GovernmentMemberDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "membershipId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "GOVERNMENT_DASHBOARD_ADMIN",
              "GOVERNMENT_DASHBOARD"
            ]
          },
          "accountStatus": {
            "type": "string",
            "enum": [
              "PENDING_ACTIVATION",
              "ACTIVE",
              "SUSPENDED"
            ]
          },
          "membershipStatus": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "accountId",
          "membershipId",
          "displayName",
          "phone",
          "role",
          "accountStatus",
          "membershipStatus"
        ]
      },
      "CreateGovernmentMemberDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "13800138000"
          }
        },
        "required": [
          "displayName",
          "phone"
        ]
      },
      "PublicAcademyCampusDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "divisionCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          },
          "studentCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "divisionCode",
          "status"
        ]
      },
      "SetPublicAcademyCampusStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      }
    }
  }
}